Fixed a bug where AzCopy was not reporting performance numbers  (#2636)
