+++ azure-pipelines.yml
@@ -165,7 +165,7 @@ jobs:
       - task: PowerShell@2
         inputs: 
           targetType: 'inline'
-          script: 'Install-Module -Name Az -Scope CurrentUser -Repository PSGallery -AllowClobber -Force'
+          script: 'Install-Module -Name Az.Accounts -Scope CurrentUser -Repository PSGallery -AllowClobber -Force'
           pwsh: 'true'
         displayName: 'Install Powershell Az Module'
       - task: AzureCLI@2
