+++ ste/sourceInfoProvider-File.go
@@ -191,9 +191,12 @@ func newFileSourceInfoProvider(jptm IJobPartTransferMgr) (ISourceInfoProvider, e
 	// due to the REST parity feature added in 2019-02-02, the File APIs are no longer backward compatible
 	// so we must use the latest SDK version to stay safe
 	//TODO: Should we do that?
+	ctx := jptm.Context()
+	ctx = withPipelineNetworkStats(ctx, nil)
+
 	return &fileSourceInfoProvider{
 		defaultRemoteSourceInfoProvider: *base,
-		ctx:                             jptm.Context(),
+		ctx:                             ctx,
 		cacheOnce:                       &sync.Once{},
 		srcShareClient:                  s.NewShareClient(jptm.Info().SrcContainer),
 		sourceURL:                       source.URL()}, nil
@@ -341,7 +344,7 @@ func (p *fileSourceInfoProvider) GetMD5(offset, count int64) ([]byte, error) {
 			return response.ContentMD5, nil
 		} else {
 			// compute md5
-			body := response.NewRetryReader(p.jptm.Context(), &file.RetryReaderOptions{MaxRetries: MaxRetryPerDownloadBody})
+			body := response.NewRetryReader(p.ctx, &file.RetryReaderOptions{MaxRetries: MaxRetryPerDownloadBody})
 			defer body.Close()
 			h := md5.New()
 			if _, err = io.Copy(h, body); err != nil {
